refactor(ask-user): own app-left inbox contribution
