+++ apps/workspace-playground/src/front/App.tsx
@@ -3,7 +3,7 @@ import { createDeckPlugin } from "@hachej/boring-deck/front"
 import type { DeckWidgetDefinition } from "@hachej/boring-deck/shared"
 import { WorkspaceProvider, INBOX_DEMO_SESSION_ID, inboxDemoPlugin } from "@hachej/boring-workspace"
 import { WorkspaceAgentFront, WorkspaceFullPagePanel, parseFullPagePanelLocation } from "@hachej/boring-workspace/app/front"
-import { askUserPlugin } from "@hachej/boring-ask-user/front"
+import { createAskUserPlugin } from "@hachej/boring-ask-user/front"
 import { SHOWCASE_SESSION_ID, seedShowcase } from "./showcaseMessages"
 
 function isShowcaseRoute(): boolean {
@@ -16,77 +16,9 @@ function isFullPageRoute(): boolean {
   return window.location.pathname === "/full-page" || window.location.pathname === "/full-page/"
 }
 
-// `?multi=1` renders a populated multi-project app-left pane with mock data, so
-// the design can be reviewed with realistic content (not an empty new-user shell).
-function isMultiRoute(): boolean {
-  if (typeof window === "undefined") return false
-  return new URLSearchParams(window.location.search).get("multi") === "1"
-}
-
 function isInboxDemoRoute(): boolean {
   if (typeof window === "undefined") return false
   return new URLSearchParams(window.location.search).get("inboxDemo") === "1"
-}
-
-const HOUR = 1000 * 60 * 60
-const MULTI_NOW = 1_782_240_000_000
-// Active project's sessions (the workspaceId below). Two are pinned.
-const MULTI_SESSIONS = [
-  { id: "m-s1", title: "Q3 revenue analysis", updatedAt: MULTI_NOW - HOUR * 0.4 },
-  { id: "m-s2", title: "Churn cohort breakdown", updatedAt: MULTI_NOW - HOUR * 3 },
-  { id: "m-s3", title: "Pricing experiment readout", updatedAt: MULTI_NOW - HOUR * 26 },
-  { id: "m-s4", title: "Onboarding funnel deep-dive", updatedAt: MULTI_NOW - HOUR * 52 },
-  { id: "m-s5", title: "Net revenue retention model", updatedAt: MULTI_NOW - HOUR * 70 },
-  { id: "m-s6", title: "Board deck — Q3 numbers", updatedAt: MULTI_NOW - HOUR * 120 },
-]
-const MULTI_PROJECTS = [
-  { id: "proj-alpha", name: "Revenue analytics", available: true, sessionCount: 6, blockedCount: 1 },
-  {
-    id: "proj-beta",
-    name: "Growth experiments",
-    available: true,
-    sessionCount: 4,
-    blockedCount: 2,
-    sessions: [
-      { id: "b1", title: "Landing page A/B test", updatedAt: MULTI_NOW - HOUR * 5 },
-      { id: "b2", title: "Referral loop model", updatedAt: MULTI_NOW - HOUR * 30 },
-      { id: "b3", title: "Activation email copy", updatedAt: MULTI_NOW - HOUR * 96 },
-      { id: "b4", title: "Paywall placement test", updatedAt: MULTI_NOW - HOUR * 140 },
-    ],
-  },
-  {
-    id: "proj-gamma",
-    name: "Data platform",
-    available: true,
-    sessionCount: 3,
-    sessions: [
-      { id: "g1", title: "dbt model refactor", updatedAt: MULTI_NOW - HOUR * 8 },
-      { id: "g2", title: "Warehouse cost audit", updatedAt: MULTI_NOW - HOUR * 44 },
-      { id: "g3", title: "Event schema migration", updatedAt: MULTI_NOW - HOUR * 200 },
-    ],
-  },
-  {
-    id: "proj-delta",
-    name: "Customer research",
-    available: true,
-    sessionCount: 2,
-    sessions: [
-      { id: "d1", title: "Churn interview synthesis", updatedAt: MULTI_NOW - HOUR * 18 },
-      { id: "d2", title: "NPS verbatim themes", updatedAt: MULTI_NOW - HOUR * 60 },
-    ],
-  },
-  { id: "proj-epsilon", name: "Marketing ops", available: true, sessionCount: 9 },
-]
-
-function seedMultiRouteStorage(): void {
-  if (typeof window === "undefined" || !isMultiRoute()) return
-  try {
-    // Two pinned sessions from the active project + one pinned project.
-    window.localStorage.setItem("boring-workspace:pinned-sessions:proj-alpha", JSON.stringify({ ids: ["m-s2", "m-s6"] }))
-    window.localStorage.setItem("boring-workspace:pinned-projects", JSON.stringify({ ids: ["proj-gamma"] }))
-  } catch {
-    // ignore storage failures in the mock
-  }
 }
 
 interface WorkspaceMeta {
@@ -114,6 +46,7 @@ const playgroundDeckPlugin = createDeckPlugin({
   },
 })
 
+const askUserPlugin = createAskUserPlugin({ appLeftInbox: true })
 const baseWorkspacePlugins = [askUserPlugin, playgroundDeckPlugin]
 const externalPluginsEnabled = (import.meta as ImportMeta & { env?: Record<string, string> }).env?.VITE_BORING_EXTERNAL_PLUGINS === "1"
 
@@ -168,10 +101,8 @@ function WorkspaceFullPageShell() {
 
 export function WorkspaceShell() {
   resetPlaygroundStorageIfRequested()
-  seedMultiRouteStorage()
   const showcase = useMemo(isShowcaseRoute, [])
   const fullPage = useMemo(isFullPageRoute, [])
-  const multi = useMemo(isMultiRoute, [])
   const inboxDemo = useMemo(isInboxDemoRoute, [])
   const workspacePlugins = useMemo(() => inboxDemo ? [...baseWorkspacePlugins, inboxDemoPlugin] : baseWorkspacePlugins, [inboxDemo])
   const [projectName, setProjectName] = useState("Workspace")
@@ -235,32 +166,6 @@ export function WorkspaceShell() {
     return <WorkspaceFullPageShell />
   }
 
-  if (multi) {
-    return (
-      <WorkspaceAgentFront
-        workspaceId="proj-alpha"
-        apiBaseUrl=""
-        workspaceLayout="plugin-tabs"
-        appLeftLayoutMode="multi-project"
-        appTitle="Acme Analytics"
-        workspaceLabel="Revenue analytics"
-        workspaceSectionTitle="Projects"
-        provisionWorkspace={false}
-        sessions={MULTI_SESSIONS}
-        appLeftProjects={MULTI_PROJECTS}
-        appLeftActiveProjectId="proj-alpha"
-        onSwitchAppLeftProject={() => {}}
-        onOpenAppLeftProjectSession={() => {}}
-        onShowMoreAppLeftProjectSessions={() => {}}
-        onCreateAppLeftProject={() => {}}
-        onOpenAppLeftProjectSettings={() => {}}
-        onOpenAppLeftProjectInNewTab={() => {}}
-        plugins={workspacePlugins}
-        chatParams={{ thinkingControl: true }}
-      />
-    )
-  }
-
   if (!metaLoaded) {
     return <div className="h-screen w-screen bg-background" />
   }
@@ -269,11 +174,9 @@ export function WorkspaceShell() {
     <WorkspaceAgentFront
       workspaceId={inboxDemo ? "playground-inbox-demo" : showcase ? "playground" : workspaceId}
       apiBaseUrl=""
-      apiTimeout={60_000}
       persistenceEnabled
       providerStorageKey={inboxDemo ? "boring-ui-v2:layout:playground:inbox-demo" : showcase ? "boring-ui-v2:layout:playground" : `boring-ui-v2:layout:playground:${workspaceId}`}
       workspaceLayout="plugin-tabs"
-      appLeftHeaderMode="hidden"
       appTitle={showcase ? "Boring" : projectName}
       workspaceLabel={showcase ? undefined : projectName}
       defaultSessionTitle={showcase ? "New session" : projectName}
