--- packages/workspace/src/app/front/WorkspaceInboxHost.tsx
@@ -1,4 +1,0 @@
-export {
-  useWorkspaceShellCapabilitiesHost as useWorkspaceInboxHost,
-  type WorkspaceShellCapabilitiesHostResult as WorkspaceInboxHostResult,
-} from "./WorkspaceShellCapabilitiesHost"
