--- packages/workspace/src/app/front/useWorkspaceInboxShellController.ts
@@ -1,1 +1,0 @@
-export { useWorkspaceShellCapabilitiesController as useWorkspaceInboxShellController } from "./useWorkspaceShellCapabilitiesController"
