+++ packages/workspace/src/app/front/workspaceBuiltinPlugins.ts
@@ -1,5 +1,4 @@
 import { filesystemPlugin } from "../../plugins/filesystemPlugin/front"
-import { workspaceInboxPlugin } from "../../plugins/inboxPlugin/front"
 import { captureBootstrapPlugins } from "../../shared/plugins/bootstrap"
 import { PluginError } from "../../shared/plugins/errors"
 import type { BoringFrontFactoryWithId, CapturedFrontPlugin } from "../../shared/plugins/frontFactory"
@@ -8,23 +7,15 @@ const RESERVED_APP_LEFT_ACTION_IDS = new Set(["plugins", "skills"])
 
 export interface CaptureWorkspaceFrontPluginsOptions {
   plugins?: BoringFrontFactoryWithId[]
-  inboxEnabled: boolean
   excludeDefaults?: string[]
 }
 
 export function captureWorkspaceFrontPlugins({
   plugins,
-  inboxEnabled,
   excludeDefaults,
 }: CaptureWorkspaceFrontPluginsOptions): CapturedFrontPlugin[] {
-  const providedPlugins = inboxEnabled
-    ? plugins ?? []
-    : (plugins ?? []).filter((plugin) => plugin.pluginId !== workspaceInboxPlugin.pluginId)
-  const pluginsWithInbox = inboxEnabled && !providedPlugins.some((plugin) => plugin.pluginId === workspaceInboxPlugin.pluginId)
-    ? [workspaceInboxPlugin, ...providedPlugins]
-    : providedPlugins
   const defaultPlugins = (excludeDefaults ?? []).includes(filesystemPlugin.pluginId) ? [] : [filesystemPlugin]
-  const captured = captureBootstrapPlugins({ plugins: pluginsWithInbox, defaults: defaultPlugins, excludeDefaults })
+  const captured = captureBootstrapPlugins({ plugins: plugins ?? [], defaults: defaultPlugins, excludeDefaults })
 
   for (const plugin of captured) {
     for (const action of plugin.registrations.appLeftActions) {
